+++ playground/src/app.config.js
@@ -20,7 +20,6 @@ const appConfig = {
   drawer: {
     content: DrawerContent,
     logo: DrawerLogo,
-    elements: [],
   },
 };