+++ playground/src/index.tsx
@@ -1,5 +1,4 @@
 import "assets/styles/styles.output.css";
-import theme from "theme.config.json";
 
 if ("serviceWorker" in navigator && process.env.NODE_ENV === "production") {
   navigator.serviceWorker
@@ -20,7 +19,6 @@ import("modular-preview").then(({ initApplication }) => {
   import("engine.config").then(({ default: engine }) => {
     initApplication({
       engine,
-      theme,
       onComplete: (App) => {
         import("react-dom").then(({ render }) => {
           render(App, document.getElementById("root"));