+++ src/app/constants/default-configs.ts
@@ -1,0 +1,28 @@
+export const defaultTheme = {
+  header: {
+    style: {
+      height: "20%",
+      borderBottom: "2px solid #c0c0c0",
+    },
+  },
+  footer: {
+    style: {
+      height: "10%",
+      borderTop: "2px solid #c0c0c0",
+    },
+  },
+  router: {
+    style: { height: "70%" },
+  },
+  drawer: {
+    style: {
+      borderRight: "2px solid #c0c0c0",
+    },
+  },
+};
+
+export const defaultEngineConfig = {
+  redux: { ui: true, modal: true, epics: [], reducers: {}, preload: {} },
+};
+
+export const defaultAppConfig = {};